requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/base.in
#
asgiref==3.11.0
    # via django
django==6.0
    # via
    #   -r requirements/base.in
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements/base.in
psycopg[binary]==3.3.2
    # via -r requirements/base.in
psycopg-binary==3.3.2
    # via psycopg
python-dotenv==1.2.1
    # via -r requirements/base.in
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via psycopg
